Merge HttpConnectorMultiplexer into HttpDownloader.
Merge remaining useful test cases into HttpDownloaderTest.

Remove RetryingInputStream as the HttpConnector class already handles
retries for HTTP status codes indicating temporary failures.

This means we will not retry and resume downloads anymore when a TCP
connection fails in the middle of a download (e.g. due to a network
outage), but I think this case is rare enough that we don't need the
added complexity in our code.